Crate phy

The physics bundle.

This bundle provides orbital logic and collision event dispatching.

Modules

collision
Handles collision of bounding boxes and maintains OrbitIndexs.

Structs

BodyIndex
A data structure to store a set of bodies.

Orbit
Represents a Keplerian orbit.

OrbitalState
Represents the ECI position and velocity of an orbit at time t.

Star
An object with a non-negligible gravitational field.

Enums

Body
A gravity-dependent object in a solar system.