Enum phy::Body

pub enum Body {
    Root(RootBody),
    Orbiting(OrbitingBody),
    Accelerating(AccelBody),
    Standing(StandingBody),
}

A gravity-dependent object in a solar system.

Variants

Root(RootBody)

The root star, which has no parent star

Orbiting(OrbitingBody)

An orbiting object.

Accelerating(AccelBody)

An object accelerating due to forces other than gravitation.

Standing(StandingBody)

An object standing on a star.

Implementations

impl Body

pub fn parent(&self) -> Option<Entity>

Retrieves the parent of this body.

Returns None if this body is a root star.

pub fn position(&self, t: GameInstant) -> Position

Computes the position of the entity.

Panics if the body is a root star.

pub fn velocity(&self, t: GameInstant, mass: Mass) -> Velocity

Computes the velocity of this body.

Panics if the body is a root star.