Struct phy::Orbit

pub struct Orbit { /* fields omitted */ }

Represents a Keplerian orbit.

This struct represents the trajectory of a Keplerian orbit in a star system. This is isomorphic to an OrbitalState given a specific mass and time frame.

Extra data are stored in this struct for efficient computation.

Implementations

impl Orbit

pub fn from_states(state: OrbitalState, t: GameInstant, m: Mass) -> Self

Computes the Orbit parameters at a specific OrbitalState.

pub fn approx_position(&self, t: GameInstant, tolerance: Length) -> Position

Approximate the position of the orbit at time t.

The tolerance is measured on each component, and the modulus would have sqrt(2) times of this tolerance.

pub fn approx_velocity(
&self,
t: GameInstant,
m: Mass,
tolerance: Length
) -> Velocity

Approximate the velocity of the orbit at time t.

The tolerance is measured on each component, and the modulus would have sqrt(2) times of this tolerance.

pub fn approx_bearing(
&self,
t: GameInstant,
m: Mass,
tolerance: Theta
) -> Bearing

Approximate the bearing of the orbit at time t.

pub fn approx_radius(
&self,
t: GameInstant,
m: Mass,
tolerance: Length
) -> Length

Approximate the distance of the orbit from origin at time t.

pub fn radius_comparator(
&self,
radius: Length
) -> impl Fn(GameInstant) -> bool + 'static

Returns an efficient function to determine whether the orbit has radius greater than radius at arbitrary time.

pub fn compare_radius(
&self,
other: &Self,
t: GameInstant,
tolerance: Length
) -> Ordering

Compare which orbit has radius.

Ordering::Greater implies that other has a shorter arc to self along the clockwise direction than along the counterclockwise direction.

If the difference between the two radii is less than floating point precision in internal operations, Ordering::Equal may be returned.

pub fn compare_bearing(
&self,
other: &Self,
t: GameInstant,
tolerance: Length
) -> Ordering

Compare which orbit has greater bearing along the minor arc.

Ordering::Greater implies that self has a shorter arc to other along the clockwise direction than along the counterclockwise direction.

If the difference between the two arcs (in the mod-2pi field) is less than floating point precision in internal operations, Ordering::Equal may be returned.

pub fn bearing_in_range(
&self,
low: Bearing,
high: Bearing,
t: GameInstant
) -> impl Fn(GameInstant) -> bool

Tests whether the bearing is in the arc starting from low, extending counterclockwise until high.

pub fn when_intersect_radius(
&self,
radius: Length,
after: GameInstant
) -> Option<GameInstant>

Computes the time (starting from after) when this orbit intersects with the circle of radius radius (either moving into the circle or out of it).

This method only computes a lower bound, i.e. from after up to the returned value, the orbit must not cross the radius circle. This method is not guaranteed to return the ideal lower bound, but is designed to converge towards the intersection efficiently.

This method returns None if the orbit can never, or not in computationally relevant time, intersect with radius.

pub fn when_intersect_bearing(
&self,
other: &Self,
delta: Theta,
after: GameInstant
) -> Option<GameInstant>

Computes the time (starting from after) when the bearings of the two orbits have less than delta apart.

This method only computes a lower bound. It returns Some(after) if the two bearings are already less than delta apart. It returns None if the orbit can never, or not in computationally relevant time, reach less than delta bearing difference.

pub fn bearing_to_ta(&self, bearing: Bearing) -> TrueAnomaly

Converts bearing to true anomaly.

pub fn ta_to_ea(&self, ta: TrueAnomaly) -> EccenAnomaly

Converts true anomaly to eccentric anomaly.

pub fn ea_to_ta(&self, ea: EccenAnomaly) -> TrueAnomaly

Converts eccentric anomaly to true anomaly.

pub fn ea_to_ma(&self, ea: EccenAnomaly) -> MeanAnomaly

Converts eccentric anomaly to mean anomaly.